wizardlm2-function-calling/requirements.txt
transformers
openai
langchain
torch
langchain_openai
python-dotenv